fix a bug with local 2erisc sync in 2erisc fabric mode (#31595)
This resolves a read after write hazard. When we do atomic increment on
write with the stream hardware in the local erisc sync method, we are
writing to one register and reading back the result from another
register. Further there can be a delay (atleast one cycle) here for the
value to update. The new implementation uses overlay scratch registers
that are written to and read from directly (i.e. same address for both)
so the reads and writes are serialized from the risc perspective.


### Ticket
Closes #31250 

### Checklist
- [x] All post commit - T3K Quick:
https://github.com/tenstorrent/tt-metal/actions/runs/18980649736
- [x] Blackhole Post commit:
https://github.com/tenstorrent/tt-metal/actions/runs/18980667344
- [x] BH Multicard Unit Test:
https://github.com/tenstorrent/tt-metal/actions/runs/18980559097
- new test run here:
https://github.com/tenstorrent/tt-metal/actions/runs/18980559097/attempts/1